test(office): organize decision idempotency regressions
