+++ apps/backend/internal/office/dashboard/agent_decisions_test.go
@@ -145,51 +145,6 @@ func TestRecordAgentDecision_RejectedQueuesAssigneeRunWithReason(t *testing.T) {
 	}
 }
 
-// TestRecordAgentDecision_TwoRejectionRoundsQueueDistinctRuns is the
-// agent-path counterpart to TestRequestTaskChanges_TwoRoundsQueueDistinctRuns:
-// two rejected verdicts against the same task within the idempotency window
-// each queue their own assignee run, keyed off the decision that produced
-// them rather than the static (reason, task, agent) tuple.
-func TestRecordAgentDecision_TwoRejectionRoundsQueueDistinctRuns(t *testing.T) {
-	deps := newTestDeps(t)
-	insertTestTaskWithAssignee(t, deps.db, "rej2", "ws-d", "REJ2", "in_review", 2, "asg-rej2")
-	mustAddParticipant(t, deps, "rej2", "agent-rev", models.ParticipantRoleReviewer)
-
-	q := &stubApprovalQueuer{}
-	deps.svc.SetApprovalReactivityQueuer(q)
-
-	first, err := deps.svc.RecordAgentDecision(context.Background(), dashboard.RecordAgentDecisionInput{
-		TaskID:         "rej2",
-		AgentProfileID: "agent-rev",
-		Decision:       engine.DecisionRejected,
-		Reason:         "round one",
-	})
-	if err != nil {
-		t.Fatalf("first RecordAgentDecision: %v", err)
-	}
-	second, err := deps.svc.RecordAgentDecision(context.Background(), dashboard.RecordAgentDecisionInput{
-		TaskID:         "rej2",
-		AgentProfileID: "agent-rev",
-		Decision:       engine.DecisionRejected,
-		Reason:         "round two",
-	})
-	if err != nil {
-		t.Fatalf("second RecordAgentDecision: %v", err)
-	}
-	if len(q.runs) != 2 {
-		t.Fatalf("runs = %d, want 2: %#v", len(q.runs), q.runs)
-	}
-	if first.DecisionID == second.DecisionID {
-		t.Fatalf("engine returned identical DecisionIDs for two separate decisions: %q", first.DecisionID)
-	}
-	if want := "decision:" + first.DecisionID; q.runs[0].IdempotencyKey != want {
-		t.Errorf("runs[0].IdempotencyKey = %q, want %q", q.runs[0].IdempotencyKey, want)
-	}
-	if want := "decision:" + second.DecisionID; q.runs[1].IdempotencyKey != want {
-		t.Errorf("runs[1].IdempotencyKey = %q, want %q", q.runs[1].IdempotencyKey, want)
-	}
-}
-
 // insertTestTaskNoStep inserts a task row with no workflow_step_id bound
 // (the schema default is ” — see tasks.workflow_step_id NOT NULL DEFAULT
 // ”), exercising the AC-7/AC-55(1) precondition RecordAgentDecision must
